add font-awesome icon configuration
Created a `config` directory that we can use for future config. The pattern
we'd like to use is create file that exports an object with a `configure`
function. Then in `main.js` all the included config files will have configure
called on them.

Getting the icons working was tricky but this [Github issue comment](https://github.com/storybookjs/storybook/issues/12432#issuecomment-785340569)
saved the day