configure media query mixins for easier responsive styling
CSS variables are [not supported in media queries](https://www.w3.org/TR/css-variables-1/#using-variables)
([this Stack Overflow](https://stackoverflow.com/a/40723269) is more helpful).
E.g. you can't do
```
@media screen and (max-width: --phone-width) {
}
```

Sprinkling `576px` throughout the code for responsive styling felt icky,
so this is my attempt at an easier solution for us to use going forward
(easier to use, wasn't so easy to configure 😂)

We have some sass variables in `screen.scss` that we're using in `mixins.scss`
This is to just set up all the different pre-configured media queries
that we'll want to use the mixin for. Note: I added some additional breakpoints
just in case we want to do more than phone and not-phone going forward.

Then I configured our sass loader to preload the `mixins.scss` globally
so that we don't need to remember to import it into every component that we
want to responsive style. It only loads the mixins file, not all our styles
so I'm not too worried about style bloat.

If we want more mixins for reusable styles going forward, we can add them
to `mixins.scss` and it should "just work".

Here's some documentation that helped me get to this solution
- [Vue config CSS loader options](https://cli.vuejs.org/config/#css-loaderoptions)
- [Sass loader options: `additionalData`](https://github.com/webpack-contrib/sass-loader#additionaldata)
- [Getting these options to work with Storybook](https://github.com/storybookjs/storybook/issues/6743)