Updated tailwind config to search *.vue files for CSS to purge
In order to keep our production builds small, we want to purge unused
CSS from the build. PurgeCSS is included in Tailwind by default so we
updated the configuration to look at *.vue files to find what CSS classes
are used.

For info on how to write purgeable HTML, please see [the docs](https://tailwindcss.com/docs/optimizing-for-production#writing-purgeable-html)